Ignorando arquivos

+++ main.py
@@ -20,7 +20,9 @@ def verificarCadastroProduto(produtos, nome):
             return [True, chave]
 
 def primeiraInicialização():
-    open('produtos.json', 'w').write('{}')
+    if not isfile('produtos.json'): open('produtos.json', 'w').write('{}')
+    if not isfile('vendas.json'): open('vendas.json', 'w').write('[]')
+    if not isfile('caixa.json'): open('caixa.json', 'w').write('{"status": false, "historico": [], "valor": 0}')
 
 def cadastrarProduto():
     produtos = abrirJson('produtos.json')
@@ -38,39 +40,96 @@ def cadastrarProduto():
         editarJson('produtos.json', produtos)
         print('Produto Cadastrado Com Sucesso!')
 
-def vender():
-    produtosEstoque = abrirJson('produtos.json')
-    produtosAVender = list()
+class Caixa:
+    def __init__(self):
+        self.caixa = abrirJson('caixa.json')
 
-    while True:
-        produtoAVender = dict()
-        mostrarProdutos(produtosEstoque)
-    
+    def vender(self):
+        def salvarVenda(self):
+            from datetime import datetime
+
+            vendas = abrirJson('vendas.json')
+
+            dataVenda = dict()
+            dataVenda['ano'] = datetime.now().year
+            dataVenda['mes'] = datetime.now().month
+            dataVenda['dia'] = datetime.now().day
+            dataVenda['hora'] = datetime.now().hour
+            dataVenda['minuto'] = datetime.now().minute
+            dataVenda['segundo'] = datetime.now().second
+            venda = list()
+            venda.append(produtosAVender)
+            venda.append(dataVenda)
+
+            vendas.append(venda)
+
+            editarJson('vendas.json', vendas)
+
+        def modificarEstoque(self):
+            for produto in produtosAVender:
+                produtosEstoque[produto['idProdutoAVender']]['quantidade'] = produtosEstoque[produto['idProdutoAVender']]['quantidade'] - produto['quantidadeAVender']
+                editarJson('produtos.json', produtosEstoque)
+
+            
+        produtosEstoque = abrirJson('produtos.json')
+        produtosAVender = list()
+
         while True:
-            produtoAVender['nomeProdutoAVender'] = str(input('Qual produto adicionará?\n'))
-            idProdutoAVender = verificarCadastroProduto(produtosEstoque, produtoAVender['nomeProdutoAVender'])[-1]
-            if idProdutoAVender:
-                break
-            else: print('Produto Inválido!')
-
-        produtoAVender['idProduto'] = idProdutoAVender
-        produtoAVender['quantidadeAVender'] = receberInteiro('Quantidade: ')
-        produtosAVender.append(produtoAVender)
+            produtoAVender = dict()
+            mostrarProdutos(produtosEstoque)
         
-        escolhaVenda = str(input('Deseja adicionar mais produtos? ')).strip().lower()
+            while True:
+                produtoAVender['nomeProdutoAVender'] = str(input('Qual produto adicionará?\n'))
+                idProdutoAVender = verificarCadastroProduto(produtosEstoque, produtoAVender['nomeProdutoAVender'])[-1]
+                if idProdutoAVender: break
+                else: print('Produto Inválido!')
 
-        if 's' not in escolhaVenda:
-            break
-    total = float()
+            produtoAVender['idProdutoAVender'] = idProdutoAVender
+            produtoAVender['quantidadeAVender'] = receberInteiro('Quantidade: ')
+            produtosAVender.append(produtoAVender)
+            
+            escolhaVenda = str(input('Deseja adicionar mais produtos? ')).strip().lower()
 
-    for produto in produtosAVender:
-        nomeProduto = produtosEstoque[produto["idProduto"]]['nome']
-        precoProduto = produtosEstoque[produto['idProduto']]['precoDeVenda']
-        print(f'Produto: {nomeProduto} {("-" * (40 - len(nomeProduto)))} | Preço: {precoProduto} {"-" * 36} | Quantidade: {produto["quantidadeAVender"]} {"-" * 38}')
-        total += precoProduto * produto['quantidadeAVender']
-    
-    print(f'Total: {total}')
+            if 's' not in escolhaVenda: break
+        total = float()
 
+        for produto in produtosAVender:
+            nomeProduto = produtosEstoque[produto["idProdutoAVender"]]['nome']
+            precoProduto = produtosEstoque[produto['idProdutoAVender']]['precoDeVenda']
+            print(f'Produto: {nomeProduto} {("-" * (40 - len(nomeProduto)))} | Preço: {precoProduto} {"-" * 36} | Quantidade: {produto["quantidadeAVender"]} {"-" * 38}')
+            total += precoProduto * produto['quantidadeAVender']
+        
+        print(f'Total: {total}')
+
+        escolha = receberInteiro('O que deseja fazer? [ 1 ] Finalizar Venda | [ 0 ] Cancelar Venda')
+
+        if escolha == 1:
+            self.vender().modificarEstoque()
+            self.vender().salvarVenda()
+            print('Venda Finalizada')
+
+    def abrirFecharCaixa(self, status=True):
+        self.caixa['status'] = status
+        editarJson('caixa.json', self.caixa)
+
+    def verificarCaixaAbertoFechado(self):
+        if self.caixa['status']: return True
+
+    def inserirRetirarValor(self):
+        escolha = receberInteiro('O que deseja fazer? [ 1 ] Inserir [ 2 ] Retirar [ 0 ]')
+
+        if escolha == 1:
+            quantidadeAInserir = receberInteiro('Quantidade a retirar:\n')
+            self.caixa['historico'].append({"acao": "entrada", "quantidade": quantidadeAInserir})
+        elif escolha == 2:
+            quantidadeARetirar = receberInteiro('Quantidade a retirar:\n')
+            self.caixa['historico'].append({"acao": "retirada", "quantidade": -quantidadeARetirar})
+
+        editarJson('caixa.json', self.caixa)
+
+    def inserirValorVenda(self, valor):
+        self.caixa['historico'].append({"acao": "venda", "quantidade": valor})
+        
 
 def editarProduto():
     def mostrarProdutoOpcoes(produto):
@@ -114,8 +173,7 @@ def editarProduto():
             input('pressione enter')
 
 if __name__ == '__main__':
-    if not isfile('produtos.json'):
-        primeiraInicialização()
+    primeiraInicialização()
 
     while True:
         escolha = int(input('O que quer fazer? [0]Sair [1]Venda [2]Cadastrar Produto [3]Editar Produto\n'))
@@ -123,7 +181,15 @@ if __name__ == '__main__':
         if escolha == 0:
             break
         elif escolha == 1:
-            vender()
+            caixa = Caixa()
+
+            if not caixa.verificarCaixaAbertoFechado():
+                escolha = str(input('O caixa está fechado, deseja abri-lo? [ S / N ]')).strip().lower()
+
+                if 's' in escolha: caixa.abrirFecharCaixa()
+            elif caixa.verificarCaixaAbertoFechado():
+                caixa.vender()
+    
         elif escolha == 2:
             cadastrarProduto()
         elif escolha == 3:
